feat: reorganize TimeTab component and its types into a new directory structure
