+++ src/renderer/components/states/hover/components/HoverForm.tsx
@@ -27,7 +27,7 @@
 import type { ReactElement } from 'react';
 import type { useHover } from '../hooks/useHover';
 import { NAV_DOTS } from '../config/hoverConfig';
-import { TimeTab } from '../pages/TimeTab';
+import { TimeTab } from '../pages/time';
 import { LyricsTab } from '../pages/lyric';
 import { WeatherTab } from '../pages/WeatherTab';
